+++ lib/react_on_rails/test_helper/ensure_assets_compiled.rb
@@ -37,6 +37,8 @@ module ReactOnRails
         # All done if no stale files!
         return if stale_gen_files.empty?
 
+        ReactOnRails::PacksGenerator.generate
+
         # Inform the developer that we're ensuring gen assets are ready.
         puts_start_compile_check_message(stale_gen_files)