+++ spec/react_on_rails/fixtures/automated_packs_generation/components/ComponentWithServerOnly/ror_components/ComponentWithServerOnly.server.jsx
@@ -1,0 +1,1 @@
+// Empty Test Component